ci: adding pull_request triggers to all workflows (#368)
The `on.push` only considers the last push and does not consider all changes in a respective PR. This allowed for cases where not all checks are running on the last commit of a pull request, which leads to problems with the expected pr status checks.

This PR also removed a outdated licences workflow that is not needed anymore after the switch to a toml file.